human_pose_estimation/single_camera/videopose3d/requirements.txt
numpy
opencv-python
matplotlib
tqdm
pyyaml
pandas
scipy
torch
torchvision
torchaudio
